test/requirements.txt
pyserial>=3.5
git+https://github.com/simplefoc/pysimplefoc

